requirements.txt
anyio==4.6.0
certifi==2024.8.30
charset-normalizer==3.3.2
h11==0.14.0
httpcore==1.0.6
httpx==0.27.2
idna==3.10
notion-client==2.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
requests==2.32.3
six==1.16.0
sniffio==1.3.1
urllib3==2.2.3
